changed the maximum "num_points" in P9374A to the actual device limitation 100001; added a P9374A driver that is compatible with the  the old NI-VISA driver
We found a weird bug that when running the P9374A python driver on one of the measurement computers, the gettrace function raises decoding error at specific data points (n*2481). Although this computer has all the related python packages (qcodes, pyvisa) exactly the same as the other computers where the code works fine.   After a series of debugging, I found that the problem lies in the NI-VISA environment. The NI-VISA driver on this computer was too old (2011), while the other computers have at least 2020. Upgrading the NI-VISA environment to the latest version resolved the problem. There is also a python workaround that solves decoding problem without the need of updating NI-VISA, I uploaded it just for future debugging purpose.